fix(cleanup): remove unnecessary code
